chore(resolver): move to rule from engine
