+++ src/rule/textlint-module-resolver.js
